requirements.txt
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
PyJWT==2.10.1
python-dotenv==1.1.0
azure-storage-blob==12.25.1
requests==2.32.4
fastapi
uvicorn
